dev: removed unused files
