--- Assets/Speckle Connector/dmo/dScripts/dGUI/StreamPropertyDrawersHelper.cs
@@ -1,14 +1,0 @@
-using Speckle_Connector.dmo;
-
-public static class StreamPropertyDrawersHelper {
-
-#if UNITY_EDITOR
-
-    // public static string[ ] AllStreamNames( )
-    //     {
-    //         return BabyStreamManager.PossibleStream != null ? BabyStreamManager.PossibleStream.ToArray( ) : new[ ] {"Nothing"};
-    //     }
-
-#endif
-
-}
